Move tests/common/mod.rs to rustls-test crate
Everything moves, with the exception of parts that depend on the
cargo features of the rustls crate.
