+++ Cargo.toml
@@ -14,6 +14,8 @@ members = [
   "provider-example",
   # the main library and tests
   "rustls",
+  # common code for testing the core crate
+  "rustls-test",
   # benchmarking tool
   "rustls-bench",
   # experimental post-quantum algorithm support
@@ -79,6 +81,7 @@ rcgen = { version = "0.13", features = ["pem", "aws_lc_rs"], default-features = 
 regex = "1"
 ring = "0.17"
 rsa = { version = "0.9", features = ["sha2"], default-features = false }
+rustls-test = { path = "rustls-test/" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = { version = "0.10", default-features = false }
