+++ ci-bench/Cargo.toml
@@ -15,6 +15,7 @@ fxhash = { workspace = true }
 itertools = { workspace = true }
 rayon = { workspace = true }
 rustls = { path = "../rustls", features = ["ring", "aws_lc_rs"] }
+rustls-test = { workspace = true }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 tikv-jemallocator = { workspace = true }
