+++ rustls-test/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rustls-test"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+rustls = { path = "../rustls", default-features = false, features = ["std", "tls12"] }
